+++ src/main.rs
@@ -1,3 +1,4 @@
+mod app;
 mod cache;
 mod cli;
 mod config;
@@ -8,38 +9,12 @@ mod models;
 mod tests;
 mod weather;
 
-use cli::Cli;
-use config::Config;
-use display::formatter::WeatherFormatter;
-use std::time::Duration;
-use weather::{GetWeather, GetWeatherProvider};
+use app::App;
+use errors::RustormyError;
 
-fn clear_screen() {
-    print!("\x1B[2J\x1B[1;1H\x1B[?25l");
-    std::io::Write::flush(&mut std::io::stdout()).unwrap();
-}
-
-fn main() -> Result<(), errors::RustormyError> {
-    let config = Config::new(&Cli::new())?;
-    let provider = GetWeatherProvider::new(config.provider());
-    let formatter = WeatherFormatter::new(config.clone());
-
-    loop {
-        match provider.get_weather(&config) {
-            Ok(weather) => {
-                if config.live_mode() {
-                    clear_screen();
-                }
-                formatter.display(weather);
-            }
-            Err(error) => formatter.display_error(&error),
-        }
-        if !config.live_mode() {
-            break;
-        }
-        let sleep_duration = Duration::from_secs(config.live_mode_interval());
-        std::thread::sleep(sleep_duration);
-    }
+fn main() -> Result<(), RustormyError> {
+    let mut app = App::new()?;
+    app.run();
 
     Ok(())
 }