+++ src/models.rs
@@ -2,7 +2,7 @@ use clap::ValueEnum;
 use serde_derive::{Deserialize, Serialize};
 use std::fmt::{Display, Formatter, Result as FmtResult};
 
-#[derive(Debug, Default, Clone, Copy, Serialize, Deserialize, ValueEnum)]
+#[derive(Debug, Default, Clone, Copy, Eq, PartialEq, Serialize, Deserialize, ValueEnum)]
 #[serde(rename_all = "snake_case")]
 #[clap(rename_all = "snake_case")]
 pub enum Provider {